Apply calibration in weighing forms: display and record in tonnes
- WagonRecord: Bogie1/Bogie2/Total changed from int to double (tonnes)
- WeighingFormBase: add CalibrationService, abstract ToTonnes(); live display
  and captured values now shown as F2 tonnes instead of raw ADC codes
- StaticWeighingForm: ToTonnes() uses CalibrationService.Convert(code, channel)
- DynamicWeighingForm: ToTonnes() uses CalibrationService.ConvertDynamic(code, direction)
- CalibrationService.ConvertDynamic: use StartsWith("→") to match direction strings
- MainForm: pass CalibrationService to both weighing forms

+++ Models/WagonRecord.cs
@@ -4,9 +4,9 @@ public record WagonRecord(
     int      Number,
     DateTime TrainTime,
     DateTime WagonTime,
-    int      Bogie1,
-    int      Bogie2,
+    double   Bogie1,
+    double   Bogie2,
     string   Direction)
 {
-    public int Total => Bogie1 + Bogie2;
+    public double Total => Bogie1 + Bogie2;
 }
